When updating related entities, Eloquent events such as `saved` and `updated` are NOT fired.
When using the `update()` method without retrieving the entity first, the Laravel events `updating`, `updated`, `saving` and `saved` are not fired. I noticed this when I was working on an `Event` for my application and I was wondering why the saved event was only fired when a new related entity is added, but it is not fired when an existing entity is edited.